Introduce Filter class wrapping a single element of Filtering.
