+++ connectors/sources/mysql.py
@@ -13,11 +13,7 @@ import aiomysql
 from bson import Decimal128
 
 from connectors.logger import logger
-from connectors.source import (
-    BaseDataSource,
-    advanced_rules_present,
-    extract_advanced_rules,
-)
+from connectors.source import BaseDataSource
 
 MAX_POOL_SIZE = 10
 QUERIES = {
@@ -445,9 +441,9 @@ class MySqlDataSource(BaseDataSource):
                 f"Configured databases: {inaccessible_databases} are inaccessible for user {self.configuration['user']}."
             )
 
-        if advanced_rules_present(filtering):
+        if filtering and filtering.has_advanced_rules():
             for database in databases:
-                advanced_rules = extract_advanced_rules(filtering)
+                advanced_rules = filtering.get_advanced_rules()
 
                 if database in advanced_rules:
                     database_filtering = advanced_rules.get(database, {})
